requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
APScheduler==3.10.4
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.5.0
Flask==3.0.1
frozenlist==1.4.1
gevent==23.9.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
multidict==6.0.4
packaging==23.2
pluggy==1.4.0
PyJWT==2.8.0
pymongo==4.6.1
pypng==0.20220715.0
pytest==8.0.0
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
qrcode==7.4.2
Random-Word==1.0.11
requests==2.31.0
schedule==1.2.1
six==1.16.0
slack-bolt==1.18.1
slack_sdk==3.26.2
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
